Use unsigned long for time values
The computations done on time intervals were somewhat confused. Masking
out the sign bit of a `signed long` does not result in correct interval
computations (okay, if the mask was again applied to the result of a
computation, it actually would, but this was not done). This resulted in
a situation where `WattGauge` could get stuck until `reset()` was
called. In practice, this call happens often and it would get unstuck,
merely reporting wrong values for some time.

The type of `millis()` is unsigned anyway, As It Should Be(TM).

Along with changing the type of time values to `unsigned long`, I
changed the type of `WattGauge._p`, because otherwise it would have just
stuck out like a sore thumb here:

```
    unsigned long _t[3];  /* t0, t(end-1), t(end) */
    long _p[3];           /* P(sum) in t[n] */
    unsigned long _tlast; /* latest time, even without changed data */
```